everything to recent
- menu bar (thus far) with
  search bar (usable but cant search anything)
  settings (clickable, but that's it)
  help (clickable, but that's it)
