+++ BeachNav/app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme.NoActionBar">
 
         <!--
              The API key for Google Maps-based APIs is defined as a string resource.
@@ -35,14 +35,13 @@
             android:value="@string/google_maps_key" />
 
         <activity android:name=".SplashActivity"
-            android:theme="@style/SplashTheme">
+            android:theme="@style/Theme.SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <!--<activity android:name=".MainActivity">-->
             <!--<intent-filter>-->
                 <!--<action android:name="android.intent.action.MAIN" />-->
@@ -50,16 +49,28 @@
                 <!--<category android:name="android.intent.category.LAUNCHER" />-->
             <!--</intent-filter>-->
         <!--</activity>-->
-        <activity
-            android:name=".MainActivity$MapsActivity">
-            <!--android:label="@string/title_activity_maps">-->
-
+        <activity android:name=".MapActivity"
+            android:theme="@style/CompatTheme" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+            <meta-data android:name = "android.app.default_searchable"
+                android:value=".SearchResultsActivity" />
         </activity>
+        <activity android:name=".SearchResultsActivity"
+            android:label="@string/app_name"
+            android:theme="@style/CompatTheme" >
+            <intent-filter>
+                <action android:name="android.intent.action.SEARCH"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+            </intent-filter>
+            <meta-data
+                android:name="android.app.searchable"
+                android:resource="@xml/searchable"/>
+        </activity>
+
     </application>
 
 </manifest>
