+++ BeachNav/app/src/main/res/layout/activity_maps.xml
@@ -1,59 +1,72 @@
-<LinearLayout xmlns:map="http://schemas.android.com/apk/res-auto"
+<RelativeLayout xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical">
 
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/menu_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginEnd="5dp"
+        android:layout_marginStart="5dp"
+        android:layout_marginTop="5dp"
+        android:background="?attr/colorPrimary"
+        android:elevation="4dp"
+        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
+
     <fragment xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:map="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="5"
         android:id="@+id/BN.map"
-        tools:context=".MainActivity$MapsActivity"
+        tools:context=".MapsActivity"
         android:name="com.google.android.gms.maps.SupportMapFragment"/>
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal">
+    <!--<LinearLayout-->
+        <!--android:layout_width="match_parent"-->
+        <!--android:layout_height="match_parent"-->
+        <!--android:orientation="horizontal">-->
 
-        <EditText
-            android:id="@+id/editText"
-            android:layout_width="275dp"
-            android:layout_height="55dp"
-            android:layout_weight="4"
-            android:ems="10"
-            android:inputType="text"
-            android:text="Where ya goin'?"
-            android:visibility="visible"
-            tools:text="Where do you want to go?" />
+        <!--<EditText-->
+            <!--android:id="@+id/editText"-->
+            <!--android:layout_width="275dp"-->
+            <!--android:layout_height="55dp"-->
+            <!--android:layout_weight="4"-->
+            <!--android:ems="10"-->
+            <!--android:inputType="text"-->
+            <!--android:text="Where ya goin'?"-->
+            <!--android:visibility="visible"-->
+            <!--tools:text="Where do you want to go?" />-->
 
-        <Button
-            android:id="@+id/search"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="4"
-            android:onClick="onSearch"
-            android:drawableStart="@android:drawable/ic_menu_search" />
+        <!--<Button-->
+            <!--android:id="@+id/search"-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:layout_weight="4"-->
+            <!--android:onClick="onSearch"-->
+            <!--android:drawableStart="@android:drawable/ic_menu_search" />-->
 
-        <Button
-            android:id="@+id/location"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="4"
-            android:onClick="findLocation"
-            android:drawableStart="@android:drawable/ic_menu_mylocation" />
+        <!--<Button-->
+            <!--android:id="@+id/location"-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:layout_weight="4"-->
+            <!--android:onClick="findLocation"-->
+            <!--android:drawableStart="@android:drawable/ic_menu_mylocation" />-->
 
-        <Button
-            android:id="@+id/zoom"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="4"
-            android:drawableStart="@android:drawable/ic_notification_clear_all" />
-    </LinearLayout>
+        <!--<Button-->
+            <!--android:id="@+id/zoom"-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:layout_weight="4"-->
+            <!--android:drawableStart="@android:drawable/ic_notification_clear_all" />-->
+    <!--</LinearLayout>-->
 
 
-</LinearLayout>
+</RelativeLayout>
