+++ BeachNav/app/src/main/res/values/colors.xml
@@ -1,7 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+
     <color name="colorPrimary">#3F51B5</color>
     <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="colorAccent">#FFFFFF</color>
     <color name="background">#444440</color> <!--The color used in drawable/splash-->
+    <color name="colorTrade">#FF4081</color>
 </resources>
