+++ BeachNav/app/src/main/res/values/dimens.xml
@@ -1,3 +1,5 @@
 <resources>
     <dimen name="fab_margin">16dp</dimen>
+    <dimen name="activity_horizontal_margin">20dp</dimen>
+    <dimen name="activity_vertical_margin">4dp</dimen>
 </resources>
