+++ BeachNav/app/src/main/res/values/strings.xml
@@ -2,4 +2,8 @@
     <string name="app_name">BeachNav</string>
     <string name="title_activity_maps">Map</string>
     <string name="title_activity_main">MainActivity</string>
+    <string name="menu_search">Search</string>
+    <string name="menu_search_hint">Type and Search...</string>
+    <string name="menu_settings">Settings</string>
+    <string name="menu_help">Help</string>
 </resources>
