+++ BeachNav/app/src/main/res/xml/searchable.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<searchable xmlns:android="http://schemas.android.com/apk/res/android"
+    android:label="@string/menu_search"
+    android:hint="@string/menu_search_hint"  />
